Path Generation Complete
Finished implementing A* pathfinding system so that a path is procedurally generated through level 2. Sometimes, a path through the level doesn't exist, since the path can only traverse terrain between 0.4 and 0.6 in height. This means if the player spawns on an island and attempts to generate a path, nothing will happen.